feature/videos componentization done
