+++ src/app/page.tsx
@@ -2,15 +2,15 @@ import Header from "../components/Header";
 import About from "../sections/About/page";
 import Contact from "../sections/Contact/page";
 import Hero from "../sections/Hero/page";
-import Works from "../sections/Works/page";
+import FashionFilms from "../sections/FashionFilms/page";
 
 export default function Home() {
   return (
     <main>
       <Header />
-      <Hero /> 
+      <Hero />
       <About />
-      <Works />
+      <FashionFilms />
       <Contact />
     </main>
   );
