+++ src/videos/videos.ts
@@ -1,0 +1,15 @@
+import fashionFilmAversion from "@/videos/fashion-film-aversion.mp4";
+import fashionFilmBombB from "@/videos/fashion-film-bomb-b.mp4";
+import fashionFilmBomb from "@/videos/fashion-film-bomb.mp4";
+import fashionFilmBooq from "@/videos/fashion-film-booq.mp4";
+
+const FASHIONFILMES = {
+  1: fashionFilmAversion,
+  2: fashionFilmBombB,
+  3: fashionFilmBomb,
+  4: fashionFilmBooq,
+};
+
+export default {
+  FASHIONFILMES,
+};
